codegen: move `scalar_pair_element_backend_type` from `LayoutMethods` into `TypeBuilderMethods`
